+++ resources/views/components/credit-card-collapse.blade.php
@@ -37,7 +37,7 @@
 @push('scripts')
     <script src="https://js.stripe.com/v3/"></script>
     <script>
-        const stripe = Stripe('{{ config('services.stripe.key') }}');
+        const stripe = Stripe('{{ config('payit.stripe.key') }}');
 
         const elements = stripe.elements();